# pkg/api/rest/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Workflow management module",
        "title": "CM-Cicada REST API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "latest"
    },
    "basePath": "/cicada",
    "paths": {
...
        "/importErrors": {
            "get": {
                "description": "Get the importErrors.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Workflow]"
                ],
                "summary": "Get importErrors",
                "operationId": "get-import-errors",
                "responses": {
                    "200": {
                        "description": "Successfully get the importErrors.",
                        "schema": {
                            "$ref": "#/definitions/airflow.ImportErrorCollection"
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to get the importErrors.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/workflow/{wfId}/task/{taskId}": {
            "get": {
                "description": "Get the task.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Workflow]"
                ],
                "summary": "Get Task",
                "operationId": "get-task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "DB workflow ID.",
                        "name": "wfId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the task.",
                        "name": "taskId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully get the task.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_model.Task"
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to get the task.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/workflow/{wfId}/workflowRun/{wfRunId}/range": {
            "post": {
                "description": "Clear the task Instance.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Workflow]"
                ],
                "summary": "Clear taskInstances",
                "operationId": "clear-task-instances",
                "parameters": [
                    {
                        "type": "string",
                        "description": "DB workflow ID.",
                        "name": "wfId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the wfRunId.",
                        "name": "wfRunId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Workflow content",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_model.TaskClearOption"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully clear the taskInstances.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_model.TaskInstanceReference"
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to clear the taskInstances.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/workflow/{wfId}/workflowRun/{wfRunId}/taskInstances": {
            "get": {
                "description": "Get the task Logs.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Workflow]"
                ],
                "summary": "Get taskInstances",
                "operationId": "get-task-instances",
                "parameters": [
                    {
                        "type": "string",
                        "description": "DB workflow ID.",
                        "name": "wfId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "DB workflow ID.",
                        "name": "wfRunId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully get the taskInstances.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_model.TaskInstance"
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to get the taskInstances.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "airflow.ImportError": {
            "type": "object",
            "properties": {
                "filename": {
                    "description": "The filename",
                    "type": "string"
                },
                "import_error_id": {
                    "description": "The import error ID.",
                    "type": "integer"
                },
                "stack_trace": {
                    "description": "The full stackstrace..",
                    "type": "string"
                },
                "timestamp": {
                    "description": "The time when this error was created.",
                    "type": "string"
                }
            }
        },
        "airflow.ImportErrorCollection": {
            "type": "object",
            "properties": {
                "import_errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/airflow.ImportError"
                    }
                },
                "total_entries": {
                    "description": "Count of objects in the current result set.",
                    "type": "integer"
                }
            }
        },
...
        "github_com_cloud-barista_cm-cicada_pkg_api_rest_common.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "github_com_cloud-barista_cm-cicada_pkg_api_rest_model.Task": {
            "type": "object",
            "required": [
                "id",
                "name",
                "request_body",
                "task_component"
            ],
            "properties": {
                "dependencies": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "extra": {
                    "type": "object",
                    "additionalProperties": true
                },
                "id": {
                    "type": "string"
                },
                "is_deleted_task": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "path_params": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "query_params": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "request_body": {
                    "type": "string"
                },
                "task_component": {
                    "type": "string"
                }
            }
        },
        "github_com_cloud-barista_cm-cicada_pkg_api_rest_model.TaskClearOption": {
            "type": "object",
            "properties": {
                "dryRun": {
                    "type": "boolean"
                },
                "includeDownstream": {
                    "type": "boolean"
                },
                "includeUpstream": {
                    "description": "IncludeFuture     bool     `json:\"includeFuture\"`\nIncludeParentdag  bool     `json:\"includeParentdag\"`\nIncludePast       bool     `json:\"includePast\"`\nIncludeSubdags    bool     `json:\"includeSubdags\"`",
                    "type": "boolean"
                },
                "onlyFailed": {
                    "type": "boolean"
                },
                "onlyRunning": {
                    "type": "boolean"
                },
                "resetDagRuns": {
                    "type": "boolean"
                },
                "taskIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "github_com_cloud-barista_cm-cicada_pkg_api_rest_model.TaskInstance": {
            "type": "object",
            "properties": {
                "dag_id": {
                    "type": "string"
                },
                "duration_date": {
                    "type": "number"
                },
                "end_date": {
                    "type": "string"
                },
                "execution_date": {
                    "type": "string"
                },
                "is_deleted_task": {
                    "type": "boolean"
                },
                "is_software_migration_task": {
                    "type": "boolean"
                },
                "software_migration_execution_id": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "task_id": {
                    "type": "string"
                },
                "task_name": {
                    "type": "string"
                },
                "try_number": {
                    "type": "integer"
                },
                "workflow_id": {
                    "type": "string"
                },
                "workflow_run_id": {
                    "type": "string"
                }
            }
        },
        "github_com_cloud-barista_cm-cicada_pkg_api_rest_model.TaskInstanceReference": {
            "type": "object",
            "properties": {
                "dag_id": {
                    "description": "The DAG ID.",
                    "type": "string"
                },
                "execution_date": {
                    "type": "string"
                },
                "task_id": {
                    "description": "The task ID.",
                    "type": "string"
                },
                "task_name": {
                    "type": "string"
                },
                "workflow_id": {
                    "description": "DB workflow ID.",
                    "type": "string"
                },
                "workflow_run_id": {
                    "description": "The DAG run ID.",
                    "type": "string"
                }
            }
        },
...
    }
}